probably a good time to add a confirmation on delete

+++ app/controllers/wh/content/type/index.js
@@ -98,12 +98,14 @@ export default Ember.ArrayController.extend({
 
   actions: {
     deleteItem: function (item) {
-      var id = item.get('id');
-      item.destroyRecord().then(function () {
-        window.ENV.sendBuildSignal();
-        window.ENV.deleteIndex(id, this.get('contentType.id'));
-        this.send('notify', 'success', 'Item removed!');
-      }.bind(this));
+      if (window.confirm('Are you sure you want to remove ' + item.get('data.name') + '?')) {
+        var id = item.get('id');
+        item.destroyRecord().then(function () {
+          window.ENV.sendBuildSignal();
+          window.ENV.deleteIndex(id, this.get('contentType.id'));
+          this.send('notify', 'success', 'Item removed!');
+        }.bind(this));
+      }
     },
 
     toggleShowInCms: function (control) {
